68 obsolete dynamicscheduling elide (#87)
This adds #ifdef code to remove dynamic scheduling code that has not
been used for over 10 years. It makes the container API more complex and
removing this code will help to simplify API updates needed to support
64 bit entity representations in containers that will remove limits that
presently exist for maximum grid size. Currently include/Config/conf.h
contains a #define DYNAMICSCHEDULING that turns on this code. Comment
out this line to test if your code makes use of any of this old API.
Previous versions of the lagrangian particle code made calls to the API
but did not use it. This is a preliminary step to completely removing
this API after the 4.1 stable release is created.