Cleanup of git configure system  (#89)
This cleanup modifies the build system to create the file
include/Config/version_info.h which contains automatically generated
defines that can be used to control version dependent compilation of
Loci applications. Specifically it defines LOCI_VERSION, LOCI_BRANCH,
LOCI_VERSION_MAJOR and LOCI_VERSION_MINOR. The previous implementation
defined LOCI_VERSION and LOCI_BRANCH on the command line, while this
moves it into the include files. The other two options defined the
integer value for the major and minor version number which can be used
to use macro directed compilation in Loci codes that want to be able to
compile against multiple Loci versions. In addition, this cleans up
previous code that still used the $Name: syntax that was used to pass
some of this information from the CVS version control system. This is no
longer needed as we are no longer using CVS and can directly access the
LOCI_VERSION and LOCI_BRANCH macros to perform this feature. This is one
important step that needs to be completed before creating a branch for
stabilizing beta testing.