Fixed issue with finding Loci libraries on unit tests. (#91)
The test option of make was failing on MacOS machines because the unit
test executable was not finding the Loci libraries. This patch fixes
this problem by adding defining both LD_LIBRARY_PATH and
DYLD_LIBRARY_PATH on the execution lines from the makefile. Also added a
bit more output information from the makefile to output the result of
tests.