94 limiter subdirectory refactor (#95)
Incremental step from issue
https://github.com/EdwardALuke/loci/issues/17, where limiter.loci is
broken down into limiter.loci (which has the constraint, firstOrderCell,
and periodicity rules) and limiter folder, which contains files for
barth, nodal barth, venkatakrishan, V2, zero, and no limiter limiting
rules. For now, the specialized stencil rules for Venkta-limiters will
be in each of the respective files. In future work, these may be split
off as well. This is step 1 of the
https://github.com/EdwardALuke/loci/issues/17 limiter work, which will
go:

1) Limiter Subdirectory Refactor
2) Type renames (S for Scalar, V for vect3d, M for storeVec (mixture,
vector))
3) limiter header (.lh) file to remove repeatability of variables in
files
4) Niskikawa higher-order limiter (factored similar to V/V2 and not with
the new cell2nodemax/min work in
https://github.com/EdwardALuke/loci/issues/17)
5) Vect3d uniform limiter
6) storeVec (vector) uniform limiter
7) inline limiter refactors, specifically for Barth (similar to inline
for Venkata limiters)
8) max/min remaps (this may be a larger issue with adding cell-to-node
maps) for limiters, but first for Nodal Barth

---------

Co-authored-by: Raymond L. Fontenot <raymond.fontenot@RLF-RHEL-P3660.cfdrc.com>
Co-authored-by: Ed Luke <edluke@gmail.com>