102 nishikawa higher order limiter (#107)
Incremental step (4) from issue
https://github.com/EdwardALuke/loci/issues/17, where implement
Nishikawa's higher-order limiter. Current support for the limiter will
be 2nd (which is effectively Venkta's limiter) through 5th order.
Appropriate tests within FVMModUnitTests/FVM of the limiter will be
added in a separate issue. This is step 4 of the
https://github.com/EdwardALuke/loci/issues/17 limiter work, which will
go:

1) https://github.com/EdwardALuke/loci/issues/94 Limiter Subdirectory
Refactor
2) https://github.com/EdwardALuke/loci/issues/97
3) limiter header (.lh) file to remove repeatability of variables in
files
4) Niskikawa higher-order limiter (factored similar to V/V2 and not with
the new cell2nodemax/min work in
https://github.com/EdwardALuke/loci/issues/17)
5) Vect3d uniform limiter
6) storeVec (vector) uniform limiter
7) inline limiter refactors, specifically for Barth (similar to inline
for Venkata limiters)
8) max/min remaps (this may be a larger issue with adding cell-to-node
maps) for limiters, but first for Nodal Barth

---------

Co-authored-by: Raymond L. Fontenot <raymond.fontenot@RLF-RHEL-P3660.cfdrc.com>