103 vect3d uniform limiter min directional (#109)
Incremental step (5) from issue
https://github.com/EdwardALuke/loci/issues/17, where a minimal direction
limiter for vect3d will be implemented. Rules will be activated via
constraint, and the constrained rules will set each component of
limitv3d(V) to min(.x,.y,.z) of the computed limiter. Appropriate tests
within FVMModUnitTests/FVM of the limiter will be added in a separate
issue. This is step 4 of the
https://github.com/EdwardALuke/loci/issues/17 limiter work, which will
go:

1) https://github.com/EdwardALuke/loci/issues/94 Limiter Subdirectory
Refactor
2) https://github.com/EdwardALuke/loci/issues/97
3) https://github.com/EdwardALuke/loci/issues/99
4) https://github.com/EdwardALuke/loci/issues/102 
5) Vect3d uniform limiter
6) storeVec (vector) uniform limiter
7) inline limiter refactors, specifically for Barth (similar to inline
for Venkata limiters)
8) Move inline limiter functions to FVMMod/limiter_funcs.h to remove
duplicated code in all limiters
9) max/min remaps (this may be a larger issue with adding cell-to-node
maps) for limiters, but first for Nodal Barth

---------

Co-authored-by: Raymond L. Fontenot <raymond.fontenot@RLF-RHEL-P3660.cfdrc.com>