Update to segregate versioning dependencies (#125)
The automatic versioning information was causing the entire of Loci to
recompile whenever the first change to a clean git repository was made.
This was because the LOCI_VERSION string changed to include the dirty.
However, this is only being accessed in a few places where detailed
versioning information was printed. This divides the information into
two parts, one that includes version numbers that change less frequently
and the other that includes detailed version information that frequently
changes. This makes compilation when making minor git changes more
efficient.