bug fix for vect3d nodal barth limiter (#135)
Recent changes to limiter for vect3d where the smallest limiter value is
used for all components of the velocity vector caused scheduling errors
caused by improper implementation of constraints. This fix adds the
constraints to avoid scheduling conflict errors and also adds unit tests
for the nodal barth limiter using a constant function instead of a
linearly varying function. Regression tests also need to be developed,
but this gives at least some minimal testing for limiters that are not
identically unity on linear functions. Probably these should be extended
to V2 limiters as well, but this fix is focused on the nodal barth
limiter.