Update to prevent crash when rule compiler is not found (#137)
When the rule compiler was not found, the code accessed the
uninitialized data anyway. It turns out that when memory release is
delayed into the advance phase of the loop, it could trigger this, but
in this case, ignoring the rule compiler was a satisfactory solution.
The code is modified so that it will still generate a warning when this
occurs when compiled with debug mode enabled.