Fix error in data accessor in block_hash copy code (#141)
This error has been in the code for a while, but became evident after
compiling on my Mac after a recent OS upgrade. The fix should not have
any effect on existing code as the copy constructor is not typically
used in the implementation of DStore and DMap and it is unlikely that
any external application would have directly used the block_hash type.