Fix for DEBUG diagnostics of parSampleSort (#145)
This fixes issue where there were 0 sized messages in the balance
distribution step used as part of the parSampleSort infrastructure. The
diagnostic was FATAL(len<=0) ; but this diagnostic was causing aborts on
zero data transfer computations that were not actually an error. In
addition to modifying the FATAL checks, I also added a check to not
perform zero sized MPI communication (although technically this is OK,
just a waste of resources). The code has been tested on a number of
applications that use the parSampleSort infrastructure including overset
meshes and the space filling curve partitioner.