Fixed lpp warning message (#147)
The lpp warning message from retyped variable in nodal_barth.loci was
caused because a variable was retyped from store<real_t> to store<real>.
In fixing the problem, changed the typing convention to use the V and M
conventions for vect3d and mixture array types in node_values.loci. I also moved the
vecSize(M) type to the FVM.lh header as that is a generally useful
utility that should be part of the finite volume module type
definitions. I checked that nodal barth still functioned correctly using
CHEM on stand alone cases in addition to running Loci and CHEM
quickTests.