Fixed issue with cell2node(X) being satisfied even when X doesn't exist. (#149)
Added a hasCellValues(X) parametric rule that exists over nodes when X
exists over cells.  This can be used as a constraint for the unit rules
of the reciprocal distance weighted sums so that `cell2node(X)`
interpolated values only exist if the variable X exists. Tested using
loci quicktest as well as flowPsi and chem. All tests include at
coverage of some of the cell2node interpolation facilities, although
more regressions should probably be added to the Loci quicktest to cover
cell2nodeMax(X) and similar specialized interpolation methods. This
issue was raised by ATA Engineering in some of their use cases. This has
been a problem for a while that can be observed when plotting k and w
variables when running with no turbulence model. This fix also corrects
that behavior.