Fixing tests for MPI stubb option (#157)
The configure --nompi feature did not work nicely with the test scripts
under quickTest because they assumed mpirun was used to run the tests.
This provides a script mpinorun that can be used in the script when
testing under the --nompi configuration options.

This is part of the testing that is needed as part of stabilizing the
beta release.