Fixed bug in create_cell_stencil_full and create_cell_stencil (#159)
The create_cell_stencil_full and create_cell_stencil utility functions
would segfault in some isolated partitioning cases. The fault is related
to the use of the collectSet utility function. The function was replaced
with the dist_collect_entitySet and dist_expand_entitySet functions.
There are other places in the code that use the collectSet utility that
probably should be shifted to use the dist_colllect_entitySet and
dist_expand_entitySet utilities instead. This is left to future work.

There have been changes to the quickTest infrastructure to allow the
unit tests for the full and stable stencils to be run in parallel to
reproduce the error. The fix resolves the error, the flowPsi and chem
codes also pass their tests.