collectSet depreciation (#161)
This update is replacing calls to collectSet that has been associated
with bugs with a more consistent distribute_entitySet and
dist_expand_entitySet calls. Since these changes touch many features of
the Loci scheduler the quickTest code has been extended so that tests
could be run with different options (such as --big) to make sure that
these updates are not
breaking regressions. This refactor has been extensively tested both on
a wide range of processor numbers with bounds checking and debug
including no mpi configures. Tests were also performed with the chem
code and flowPsi.

Note, there are still calls to the collectSet code in the work
replication optimization. This is rarely used at present, but might be
useful as we start to explore GPGPU optimizations. Removal of these
calls will be considered at that time.

This should probably be tested against GGFS, particularly to make sure
that the fix for the stencil generation code still functions properly.