Graceful warning for old AMR API usage (#175)
These simple changes resolve linking errors when compiling 4.0 versions
of CHEM with the dev branch and add diagnostics when executing schedule
queries for the older API. The changes have no effect on the update AMR
API, but allow for users to evaluate dev features using the current
stable CHEM release. This will be helpful for evaluating new PETSc GPGPU
support.