Adding random partition option (#183)
This is adding a new random partitioner which partitions faces, nodes,
and cells randomly. This option is useful for stress testing
communication code as it results in high communication exposing more
edge cases. As part of this I streamlined the communication of the
partition selection from Loci::Initialize to the grid reader using a
enum type rather than separate boolean for each partition type.