18 fix repeatability bug in ptscotch parallel graph partitioner interface  (#187)
This fixes issue with ptscotch where running on the same grid on the
same number of processors would produce different partitions. Some of
the issue was with configuration options for PTScotch installation.
Added information to install.txt for the correct configuration and added
code to Loci Initialization to set the random seed to a known value.