Bump submodules: HDF5=hdf5-1.14.5, PETSc=v3.22.2, Scotch=v7.0.8 (#188)
We had some gitmodules that weren't pinned to release tag commits in
their respective repositories. I updated them to point to close tags to
the commits that they were pointing to. The Scotch version was the most
important because it has new CMAKE options that were not available in
the old version that the gitmodules were using.