Update to add conditionals to loci preprocessor (#184)
It is not possible to use c preprocessor features to disable compilation
of Loci preprocessor code as lpp runs before the C preprocessor. This
makes it impossible to implement guards such as

\#if LOCI_VERSION_MAJOR <= 4 && LOCI_VERSION_MINOR <=1
\#else
\#endif

This update adds the same features to lpp allowing $if, $ifdef $ifndef
$else $endif and $define so that it will be possible to write code to
conform to changing Loci versions. It also automatically defined
LOCI_VERSION_MAJOR and LOCI_VERSION_MINOR to be used for controlling
specifications according the the version of Loci you are compiling with.