186 ptscotch configure support (#191)
This updates configure to support --with-scotch for selecting ptscotch
(although --with-metis will also work) Also added --with-parmetis option
for when parMETIS is installed in a separate directory. Fixed scotch so
that it works with either lib or lib64, Removed
references to gklib since it appears that current metis and parmetis do
not reference the .so file even though they need it to compile.  To fix RUNPATH
behavior on some systems, dummy functions that make calls to SCOTCH error printing
or METIS libraries are made that force loading these libraries when libLoci is loaded.