Dependency build script (#193)
This is a shell script that automates the configuration and build
process for the dependencies in the `ext/` folder. The goal being to
abstract away some of the pain of building all the loci dependencies for
users. This is a first pass and there are likely issue with using it on
other flavors of Linux. I put it at the top level for now, but we could
move it elsewhere if the script is deemed useful enough to include in
the repository.

It supports a `--with-metis` and `--with-parmetis` flag in a similar
fashion as loci. It has a `--prefix` option where a user can specify
where to put the install dependencies. I kept a simple
`<dependency>_build` and `<dependency>_install/ `directory structure for
the build/install process of the dependencies. There's a `--clean` flag
that basically deletes the `<dependency>_build/` folder and
`<dependency>_install` folder before re-building the dependencies. If
the `--prefix` is passed with the `--clean`, it also deletes the install
directories at the `prefix/` location.

There's a `--clean-ext` flag that basically wipes out everything that
was compiled in the `ext/ `folder and resets all the submodules to their
uninitialized state (basically the state that the repo's `ext/` folder
is in when you clone it) for any big oopsies that someone might make.

The script also prints out the installation locations with a sample loci
`configure` command and an `LD_LIBRARY_PATH` with all the relevant paths
included if a user wants to add it to their shell environment.