[#114] Change how Spack configuration is set up.
Because the setup-spack action sets the SPACK_DISABLE_LOCAL_CONFIG
variable in $GITHUB_ENV, we needed to completely rewrite how the
configuration files are found. Instead of relying on setting the
user-scope directory to the .spack subdirectory, we activate the Spack
environment first and add the needed `.yaml` files into the
environment's scope.