Fixed issue for online adaptation of small grid on many cores. (#197)
When using online adaption on a small grid, some of the weighted
partitioners would end up assigning no entities to some processors which
would result in faults due to this edge case not being handled properly.
This fixes those bugs which allows online adaption for small meshes to
complete even on larger numbers of cores.