lpp translation of petsc FVMMod interface (#200)
I converted the petsc module to use the lpp format as a process to go
through more through review of the recent changes to support GPGPUs. In
the process I found that the double blocked matrix code had a fault, and
removed some magic numbers from the implementation. Also I changed the
reporting of device information unless the user selects this option. On
my system I have found that I cannot run with more than 32 cores with
this printing enabled without segfaulting. Also, it gets stuck on this
code when the number of cores is fairly large. I think there is alot of
contention on the shared gpu resources that is causing race condition
problems in the PETSc interface. It may be that we cannot oversubscribe
the gpus too much. Please review these changes and evaluate how they
work for gpu execution.