Selector pattern parameterized (#204)
This update adds a parametric version of the selector pattern. This
pattern uses parametric rules to automate the process of generating
constraints and validating inputs used to select different methods using
constraints. As an example of how this is used the limiter
infrastructure is updated to use the new selector pattern. This only
changes the behavior of the limiter in that it will error out if the
limiter is set to an invalid selection instead of defaulting to
venkatakrishnan, which is probably a more desirable outcome.

One advantage of the new patterns is that it is now possible to create
third party add on modules that can add new limiters without needing to
change the FVMMod.

The way that this pattern is you need to create a default rule to define
the default setting of the selector (which must be of type
param<string>) Then for each selected algorithm you add a rule such as

 $rule pointwise(OUTPUT<-selector(limiter,venkatakrishnan,V)) {}

Note, this rule is not actually executed, but it causes rules to be
instantiated to manage the creation of code that will check and create a
constraint for the selector and also create validators that will check
that one and only one of the constraints are set to ~EMPTY. The first
argument of the selector parametric variable is the input variable for
the selector, the second is the selection name, and the third is an
abbreviation of the selected name. In this case when limiter is set to
either "venkatakrishnan" or "V" then the constraint
limiter_select_venkatakrishnan will be set to UNIVERSE (~EMPTY). All
other selector generated constraints will be empty.