Set prefix on PC object to enable recognition of the preconditioner options (#234)
When preconditioner options are set for a particular system of equations in the vars file, e.g. fluid=options(..., pc_type="bjacobi", sub_pc_type="jacobi", ...), the options do not have any effect on the preconditioner object. This pull request resolves this issue by setting the options prefix for the preconditioner object.