Update to the overset header files (#246)
Created an <overset> includable header file that keeps the global C++
namespace free of using:: statements. The overset.h pulls in the global
using:: statements from Loci.h to keep the original behavior unchanged
for anyone using that header file. The new <overset> can be included and
shouldn't cause issues with PETSc due to a name collision on the `Map
`type.