added feature to generate config.log at repo top level as a user reference (#247)
Many times I find myself coming back to a loci install and cursing that
I didn't put my configure command that I used into a text file. The
configure script has 1 extra function that redirects output to a
config.log file (and maintains the original output to the screen). This
file simply records the configure command that was used to configure the
build and any of the other standard output that the configure script
emits.