New VFAD (vectorized FAD) type support (#241)
This update adds the VFAD type along with several other supporting
services. This includes the addition of the vtype (a vectorized array
type that provides convenient access to vector operations, updates to
I/O facilities to manage multiple derivative inputs in vars files
including a new unit based type that allows sensitivity derivative input
to dimensional input parameters. New unit tests for vtype and VFAD.
Updates to makefile rules to support compiling VFAD modules. Also
included removal of old NO_CMATH code that will not be needed in C++17
standard we are currently adopting which added a lot of redundant
untested code to the FAD types.