Overset documentation/formatting update (#248)
I was looking over the overset module as I have been working on things
that used the module, so I went over the whole thing and added
comments/docstrings as a way of teaching myself about the details of the
rules and data types, etc. that are used in the module. Just wanted to
put this here so that it's available if there's any value in the
comments added to the module. If not, we can close this so it doesn't
clutter the PR page.