fixed triangle face loop that was accessing out of bounds (#252)
I was seeing segfaults when trying to convert a GMSH grid to vog format.
There was a for loop that was looping too far for the case when
triangles are being considered.

On line 912 of msh2vog.cc there was:


```
for(int j=0;j<4;++j) {
  face2node[fc][j] = corner[j] ;
}
```
    
Changed it to:

```
for(int j=0;j<3;++j) {
  face2node[fc][j] = corner[j] ;
}
```


And the utility started working.