Overset module documentation updates (#254)
I'm still looking at the overset module. These are some more comments
that I added to help me understand what the module is doing. I know
there's typically two camps with regards to documentation/comments where
one of the philosophies regards comments as a point of weakness as they
can obscure the code. If this is too many comments, we can just close
this. Just wanted to post it here in case it had any value to Loci
project.